+++ primefaces/src/main/java/org/primefaces/component/badge/BadgeRenderer.java
@@ -46,6 +46,7 @@ public class BadgeRenderer extends CoreRenderer {
                     .add(badge.getStyleClass())
                     .add(!valueEmpty && value.length() == 1, Badge.NO_GUTTER_CLASS)
                     .add(valueEmpty, Badge.DOT_CLASS)
+                    .add(badge.isHidden(), "ui-state-hidden")
                     .add("large".equals(size), Badge.SIZE_LARGE_CLASS)
                     .add("xlarge".equals(size), Badge.SIZE_XLARGE_CLASS)
                     .add("info".equals(severity), Badge.SEVERITY_INFO_CLASS)
@@ -69,7 +70,7 @@ public class BadgeRenderer extends CoreRenderer {
             writer.writeAttribute("style", badge.getStyle(), "style");
         }
 
-        if (!valueEmpty) {
+        if (!valueEmpty && !badge.isHidden()) {
             writer.write(value);
         }
         writer.endElement("span");